* magick/effect.c (ThresholdImage): Don't read uninitialized pixel indexes (valgrind gripe). * coders/tiff.c (CompressionSupported): Use TIFFIsCODECConfigured() to test if a codec is supported.

+++ magick/blob.c
@@ -4099,7 +4099,7 @@ MagickExport size_t WriteBlobByte(Image *image,const magick_uint8_t value)
     case StandardStream:
     case PipeStream:
       {
-        if(putc((int)value,image->blob->file) != EOF)
+        if(putc(value,image->blob->file) != EOF)
           return 1;
         return 0;
       }
